update dependencies
sdl3: latest git
libjpeg-turbo: 3.1.2->3.1.3
libpng: 1.6.51->1.6.53
zlib-ng: 2.3.1->2.3.2
liblzma: 5.8.1->5.8.2
libarchive: 3.8.3->3.8.4
simdutf: 7.6.0->7.7.1

+++ src/GUI/CBaseUIScrollView.cpp
@@ -127,12 +127,12 @@ void CBaseUIScrollView::mouse_update(bool *propagate_clicks) {
 
     const bool wasContainerBusyBeforeUpdate = this->container->isBusy();
     if(this->bBusy) {
-        const vec2 deltaToAdd = (mouse->getPos() - this->vMouseBackup2);
+        const dvec2 deltaToAdd = (mouse->getPos() - this->vMouseBackup2);
         // debugLog("+ ({})", deltaToAdd);
 
-        anim::moveQuadOut(&this->vKineticAverage.x, (f64)deltaToAdd.x,
+        anim::moveQuadOut(&this->vKineticAverage.x, deltaToAdd.x,
                           cv::ui_scrollview_kinetic_approach_time.getDouble(), true);
-        anim::moveQuadOut(&this->vKineticAverage.y, (f64)deltaToAdd.y,
+        anim::moveQuadOut(&this->vKineticAverage.y, deltaToAdd.y,
                           cv::ui_scrollview_kinetic_approach_time.getDouble(), true);
 
         this->vMouseBackup2 = mouse->getPos();
